+++ src/modules/gauging-stations/controller.js
@@ -2,12 +2,14 @@
 
 const gaugingStationService = require('../../lib/services/gauging-station-service');
 const controller = require('../../lib/controller');
+const gaugingStationsRepo = require('../../lib/connectors/repos/gauging-stations');
+const Boom = require('@hapi/boom');
 
-const getGaugingStation = async request =>
-  controller.getEntities(
-    request.params.gaugingStationId,
-    gaugingStationService.getGaugingStations
-  );
+const getGaugingStation = async request => {
+  const { stationGuid } = request.params;
+  const gaugingStation = await gaugingStationsRepo.findOne(stationGuid);
+  return gaugingStation || Boom.notFound(`Gauging station ${stationGuid} not found`);
+};
 
 const getGaugingStationByRef = async request =>
   controller.getEntities(
@@ -15,12 +17,12 @@ const getGaugingStationByRef = async request =>
     gaugingStationService.getGaugingStationByRef
   );
 
-const getGaugingStationConditionsById = async request =>
+const getGaugingStationLicencesById = async request =>
   controller.getEntities(
     request.params.gaugingStationId,
-    gaugingStationService.getGaugingStationConditionsById
+    gaugingStationService.getGaugingStationLicencesById
   );
 
 exports.getGaugingStation = getGaugingStation;
 exports.getGaugingStationByRef = getGaugingStationByRef;
-exports.getGaugingStationConditionsById = getGaugingStationConditionsById;
+exports.getGaugingStationLicencesById = getGaugingStationLicencesById;
